+++ packages/pipes/src/portal-cache/node/node-portal-cache.ts
@@ -65,11 +65,13 @@ export abstract class PortalCacheNodeJs<ImplOptions> implements PortalCache {
     query,
     logger,
     perBlockUnfinalized,
+    finalized,
   }: {
     portal: PortalClient
     query: Query
     logger: Logger
     perBlockUnfinalized?: boolean
+    finalized?: boolean
   }): PortalBlockStream<GetBlock<Q>> {
     const queryHash = await hashQuery(query)
 
@@ -102,7 +104,7 @@ export abstract class PortalCacheNodeJs<ImplOptions> implements PortalCache {
         fromBlock: cursor.number,
         parentBlockHash: cursor.hash,
       } as Q,
-      { perBlockUnfinalized },
+      { perBlockUnfinalized, finalized },
     )) {
       const finalizedHead = batch.head.finalized?.number
       // TODO add warning